rename mailing slice file and folder name
